Crate dfn_core

Re-exports

pub use api::futures::FutureResult;

pub use api::call;

pub use api::call_explicit;

pub use endpoint::over;

pub use endpoint::over_async;

pub use endpoint::over_async_explicit;

pub use endpoint::over_async_may_reject;

pub use endpoint::over_explicit;

pub use endpoint::over_init;

pub use endpoint::over_may_reject;

Modules

api

endpoint

printer

setup

stable

Macros

eprint
Emulates the default eprint! macro.

eprintln
Emulates the default eprintln! macro.

print
Emulates the default print! macro.

println
Emulates the default println! macro.

Structs

BytesS
This tells the compiler to just read/write the raw bytes over the wire

CanisterId
A type representing a canister's [PrincipalId].

FromS
This causes values to be serialized and deserialized using rusts From and Into traits. This is nice for a quick proof of concept as these types are widely supported in the rust ecosystem, but are liable to change their encoding format without warning between compiler/library versions.

Functions

bytes
This tells communication function that you'd like you communicate using raw bytes

from
This tells communication function that you'd like you communicate using FromS