Module dfn_core::api::futures

This module contains all mechanisms required to enable asynchronous programming in Rust, based on native async Rust capabilities:

the future returned by the asynchronous System API call, and

the kickstarting/waker implementations to advance top level futures on every inter-canister callback call.

Structs

RefCounted
A reference counter wrapper we use with the CallFuture. This is required, because the future we return from the call method can either have two owners (the callback closure and the canister runtime) if the underlying system call succeeded, or just one (the canister runtime) it the system call failed.

Functions

spawn
Spawns a future running concurrently with the caller.

Type Definitions

FutureResult
The result type of the CallFuture.