Module dfn_core::endpoint

Functions

bytes
This tells communication function that you'd like you communicate using raw bytes

from
This tells communication function that you'd like you communicate using FromS

over
Over allows you to create canister endpoints easily

over_async
Allows you to create asynchronous canister endpoints

over_async_bytes

over_async_bytes_may_reject

over_async_explicit

over_async_may_reject
Like over_async, but rejects the call when the function f given as argument returns an Err. If f returns an Ok, it calls reply.

over_async_may_reject_explicit

over_bytes

over_bytes_init

over_bytes_may_reject

over_bytes_result

over_explicit
Create endpoints with the incoming and outgoing encoding schemes explicitly stated

over_from

over_init
Useful for writing the canister_init function because it doesn't return anything

over_may_reject
Like over, but rejects the call when the function f given as argument returns an Err. If f returns an Ok, it calls reply.